+++ miden-crypto/src/merkle/smt/large/storage/mod.rs
@@ -21,7 +21,7 @@ mod memory;
 pub use memory::MemoryStorage;
 
 mod updates;
-pub use updates::{StorageUpdateParts, StorageUpdates};
+pub use updates::{StorageUpdateParts, StorageUpdates, SubtreeUpdate};
 
 /// Sparse Merkle Tree storage backend.
 ///